delete function now works

+++ server/db/migrations/20241011162251_create_posts_table.js
@@ -9,6 +9,7 @@ exports.up = function (knex) {
         table.text('body').notNullable();
         table.integer('user_id').notNullable();
         table.foreign('user_id').references('id').inTable('users');
+        table.timestamps(true, true);
     });
 };
 
@@ -17,5 +18,9 @@ exports.up = function (knex) {
  * @returns { Promise<void> }
  */
 exports.down = function (knex) {
-    return knex.schema.dropTable('posts');
+    return knex.schema.hasTable('posts').then((exists) => {
+        if (exists) {
+            return knex.schema.dropTable('posts');
+        }
+    });
 };
